Updates on the bpmn stats rest
