--- components/bpmn/org.wso2.carbon.bpmn.stats.rest/src/main/webapp/WEB-INF/beans.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-  ~ Copyright 2011-2012 WSO2, Inc. (http://wso2.com)
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:jaxrs="http://cxf.apache.org/jaxrs"
-       xmlns:context="http://www.springframework.org/schema/context"
-       xsi:schemaLocation="
-http://www.springframework.org/schema/beans
-http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd
-http://cxf.apache.org/jaxrs
-http://cxf.apache.org/schemas/jaxrs.xsd">
-
-    <import resource="classpath:META-INF/cxf/cxf.xml"/>
-
-    <context:property-placeholder/>
-    <context:annotation-config/>
-    <!-- <bean class="org.springframework.web.context.support.ServletContextPropertyPlaceholderConfigurer"/>
-     <bean class="org.springframework.beans.factory.config.PreferencesPlaceholderConfigurer"/>
- -->
-
-    <jaxrs:server id="statsService" address="/statsRest">
-        <jaxrs:serviceBeans>
-            <bean class="org.wso2.carbon.bpmn.stats.rest.Service.ProcessAndTaskService"/>
-            <bean class="org.wso2.carbon.bpmn.stats.rest.Service.UserService"/>
-        </jaxrs:serviceBeans>
-        <jaxrs:providers>
-            <bean class="org.wso2.carbon.bpmn.stats.rest.Security.AuthenticationHandler"/>
-            <bean class="org.codehaus.jackson.jaxrs.JacksonJsonProvider"/>
-        </jaxrs:providers>
-    </jaxrs:server>
-
-
-</beans>
