+++ components/bpmn/pom.xml
@@ -37,7 +37,6 @@
         <module>bpmn-rest</module>
         <module>bpmn-explorer-web</module>
         <module>org.wso2.carbon.bpmn.analytics.publisher</module>
-        <module>org.wso2.carbon.bpmn.stats.rest</module>
     </modules>
 
 </project>
